+++ lib/config/home.nix
@@ -5,6 +5,7 @@
     enable = true;
 
     editors.antigravity = true;
+    editors.cursor = true;
 
     theme = {
       enable = true;